+++ wyoming/satellite.py
@@ -4,6 +4,7 @@ from dataclasses import dataclass
 from .event import Event, Eventable
 
 _RUN_SATELLITE_TYPE = "run-satellite"
+_PAUSE_SATELLITE_TYPE = "pause-satellite"
 _STREAMING_STARTED_TYPE = "streaming-started"
 _STREAMING_STOPPED_TYPE = "streaming-stopped"
 _SATELLITE_CONNECTED_TYPE = "satellite-connected"
@@ -12,7 +13,7 @@ _SATELLITE_DISCONNECTED_TYPE = "satellite-disconnected"
 
 @dataclass
 class RunSatellite(Eventable):
-    """Tell satellite to start running."""
+    """Informs the satellite that the server is ready to run a pipeline."""
 
     @staticmethod
     def is_type(event_type: str) -> bool:
@@ -24,6 +25,22 @@ class RunSatellite(Eventable):
     @staticmethod
     def from_event(event: Event) -> "RunSatellite":
         return RunSatellite()
+
+
+@dataclass
+class PauseSatellite(Eventable):
+    """Informs the satellite that the server is not ready to run a pipeline."""
+
+    @staticmethod
+    def is_type(event_type: str) -> bool:
+        return event_type == _PAUSE_SATELLITE_TYPE
+
+    def event(self) -> Event:
+        return Event(type=_PAUSE_SATELLITE_TYPE)
+
+    @staticmethod
+    def from_event(event: Event) -> "PauseSatellite":
+        return PauseSatellite()
 
 
 @dataclass